Mid_term/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
filelock==3.19.1
Flask==3.1.2
fsspec==2025.10.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
pillow==11.3.0
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
